Remove route path /sign-in.
The default route will be the base url. Since route paths are "fake", the
original route of <base>/sign-in actually made the browser request that
resource from the server - and thus error. In addition, we just want the base
url to be the launching point for the app installation on a mobile device.